Split decoder into multiple files.
