+++ openlr/CMakeLists.txt
@@ -10,6 +10,13 @@ set(
   openlr_simple_decoder.hpp
   openlr_simple_parser.cpp
   openlr_simple_parser.hpp
+  road_info_getter.cpp
+  road_info_getter.hpp
+  road_type_checkers.cpp
+  road_type_checkers.hpp
+  router.cpp
+  router.hpp
+  way_point.hpp
   )
 
 add_library(${PROJECT_NAME} ${SRC})
