+++ openlr/openlr.pro
@@ -12,9 +12,16 @@ SOURCES += \
   openlr_sample.cpp \
   openlr_simple_decoder.cpp \
   openlr_simple_parser.cpp \
+  road_info_getter.cpp \
+  road_type_checkers.cpp \
+  router.cpp \
 
 HEADERS += \
   openlr_model.hpp \
   openlr_sample.hpp \
   openlr_simple_decoder.hpp \
   openlr_simple_parser.hpp \
+  road_info_getter.hpp \
+  road_type_checkers.hpp \
+  router.hpp \
+  way_point.hpp \
